fix: move helpers to mainsail folder
