+++ packages/helpers/source/index.ts
@@ -38,7 +38,6 @@ export * from "./first-map-value.js";
 export * from "./flatten.js";
 export * from "./format-number.js";
 export * from "./format-string.js";
-export * from "./format-units.js";
 export * from "./get.js";
 export * from "./get-path-segments.js";
 export * from "./get-type.js";
@@ -130,7 +129,6 @@ export * from "./omit-by.js";
 export * from "./order-by.js";
 export * from "./ordinal.js";
 export * from "./parse.js";
-export * from "./parse-units.js";
 export * from "./parse-uri.js";
 export * from "./partition.js";
 export * from "./pascal-case.js";
