+++ packages/mainsail/source/helpers/format-units.ts
@@ -1,4 +1,4 @@
-import { ARK_MULTIPLIER, GWEI_MULTIPLIER, WEI_MULTIPLIER } from '../../mainsail/source/crypto/constants';
+import { ARK_MULTIPLIER, GWEI_MULTIPLIER, WEI_MULTIPLIER } from "../crypto/constants";
 
 export const formatUnits = (value: string, unit = 'ark'): number => {
     switch (unit.toLowerCase()) {
